Feature #29: separate string res
