+++ core/designsystem/src/main/java/com/hankki/core/designsystem/component/button/HankkiButton.kt
@@ -1,4 +1,4 @@
-package com.hankki.core.designsystem.component
+package com.hankki.core.designsystem.component.button
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
