+++ core/designsystem/src/main/java/com/hankki/core/designsystem/component/button/StoreNameSearchButton.kt
@@ -14,6 +14,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.hankki.core.designsystem.R
@@ -43,7 +44,7 @@ fun StoreNameSearchButton() {
         Spacer(modifier = Modifier.width(4.dp))
 
         Text(
-            text = "이름으로 검색",
+            text = stringResource(id = R.string.search_to_name),
             style = HankkiTheme.typography.body6,
             color = Gray400
         )
